add a warning message before deleting an item"

+++ pages/items.js
@@ -169,6 +169,7 @@ export default function Items() {
         router.push('/recipes')
       }
       if (result.isDenied) {
+        if (confirm('Warning! Deleting items could affect your recipes and calendar. Do you want to proceed?'))
         deleteItem(item)
       }
     })
